requirements.txt
google-cloud-speech==2.15.1
moviepy==1.0.3
pydub==0.25.1
pydantic==1.10.2
tqdm==4.64.1
boto3==1.26.5
requests==2.28.1
Pillow==9.2.0
fastapi==0.85.1
uvicorn==0.18.3
beautifulsoup4==4.11.1
lxml==4.9.1
openai